Real-time 3D reconstruction of dynamic scenes
The codes are a part of Point Cloud Library's KinFu and they need an
entire set up and all the additional dependencies to run. The piece of
code written between the copyright to Illinois Institute of Technology,
Chicago are written by me. This project is in collaboration with Argonne
National Laboratory. The demos show the current output that we have for
dynamic objects in the scene.